feat(sdk/privacy): make functions safer using wrapSafe
