+++ packages/sdk/src/scopes/utilities/privacy/exports.ts
@@ -1,8 +1,3 @@
-export {
-  requestPhoneAccess,
-  requestWriteAccess,
-  requestContact,
-  isRequestingWriteAccess,
-  isRequestingPhoneAccess,
-  type RequestedContact,
-} from './privacy.js'
+export { requestContact, type RequestedContact } from './requestContact.js';
+export { requestPhoneAccess, isRequestingPhoneAccess } from './requestPhoneAccess.js';
+export { requestWriteAccess, isRequestingWriteAccess } from './requestWriteAccess.js';
