+++ packages/sdk/test-utils/reset/resetPrivacy.ts
@@ -1,6 +1,7 @@
-import { resetSignal } from '@test-utils/reset.js';
+import { resetSignal } from '@test-utils/reset/reset.js';
 
-import { isRequestingPhoneAccess, isRequestingWriteAccess } from '@/scopes/utilities/privacy/privacy.js';
+import { isRequestingWriteAccess } from '@/scopes/utilities/privacy/requestWriteAccess.js';
+import { isRequestingPhoneAccess } from '@/scopes/utilities/privacy/requestPhoneAccess.js';
 
 export function resetPrivacy() {
   [isRequestingPhoneAccess, isRequestingWriteAccess].forEach(resetSignal);
